+++ src/hillel/elementary/entity/User.java
@@ -2,8 +2,8 @@ package hillel.elementary.entity;
 
 public class User extends Entity {
 
-    public User(Role role) {
-        super(role);
+    public User() {
+        super(Role.USER);
     }
 
     @Override
