+++ src/hillel/elementary/service/Service.java
@@ -4,7 +4,7 @@ import hillel.elementary.entity.Entity;
 
 import java.io.IOException;
 
-public interface Service {
+interface Service {
     public boolean checkUser(Entity entity) throws IOException;
 
     public void signUpUserData(Entity entity) throws IOException;
